Update script.js
The calculate_weekly_popularity function now returns not just the topic and count, but also the associated articles for each topic.

We added a new generate_articles_from_popularity function. This function takes the weekly popularity data and creates a list of articles based on the popularity counts.

In the main function, we added a check to see if there are any new entries. If not, we use the generate_articles_from_popularity function to create a list of articles based on the weekly popularity data.